kafka/requirements.txt
boto3==1.38.38
botocore==1.38.38
certifi==2025.6.15
charset-normalizer==3.4.2
confluent-kafka==2.10.1
idna==3.10
jmespath==1.0.1
kafka-python==2.2.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
urllib3==1.26.20
